+++ app/Http/Controllers/AdminController.php
@@ -158,6 +158,9 @@ class AdminController extends Controller
      */
     public function destroy(User $admin)
     {
+        if($admin->keranjang || $admin->favorit || $admin->alamatPengiriman->count() > 0 || $admin->nota ||$admin->pelanggan ||$admin->admin || $admin->role){
+            return redirect('/admin/admins')->with('error', 'Penghapusan data admin gagal!');
+        }
         $admin->delete();
         return redirect('/admin/admins')->with('success', 'Penghapusan data admin berhasil!');
     }
